Add audio feedback for host change in multiplayer
